Changed main.py to server.py and deleted old server.py
